Feature/30525 openzaak to connector (#69)
* feature 30525 - Created a OpenZaak Connector

Created a new OpenZaakConnector

Adjusted the ConnectorType annotation class, a new parameter is added "allowMultipleConnectors"

The parameter is stored as a boolean in a new column in the database

When a connector is stored with allowMultipleConnectors = false the createConnectorInstance() function will throw an error if a connector of the same type already exists

Cleanup all openzaak code related to creating, modifying and deleting the configuration. This is now only handled by the connector

The only existing REST method remaining is to GET the configuration. This is currently still used a lot in the front-end and unit tests. The configuration of course now comes from the connector information

Added a new liquibase customChange script that converts an open_zaak_config configuration to a connector_instance

Fixes the ConnectorDeploymentService, it will now also store an updated connector