feature Expansion Panel (#396)
* feature Expansion Panel

The Expansion panel is a Carbon Expandable Tile but with a small tweak. The boolean property 'hideAboveFoldWhenExpanded' lets you control what happens when the tile is expanded. The default value false results in the normal operation of the expandable tile. When the value is true the collapsed panel is hidden and only the expanded panel is shown

* feature Expansion Panel - Code review findings

* Update expansion-panel.component.ts

Added empty line after copyright