Rewrote the loading in of the .env.properties (#1461)
Rewrote the loading in of the .env.properties so that it makes use of the `java.util.Properties()` class, instead of manually parsing of the file. This also resolves an issue when there is a `=` sign in the value of the property. (`String.tokenize()` will split on the `=` sign, so if there are multiple `=` signs on one line, everything after the second `=` will be ignored.)